fix: move main functions into __init__.py to fix import problem
